feat: raw btn and validate style
